requirements.txt
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
flake8==7.3.0
idna==3.11
loguru==0.7.3
mccabe==0.7.0
peewee==3.18.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
redis==7.0.1
requests==2.32.5
urllib3==2.5.0
wemake-python-styleguide==1.4.0
win32_setctime==1.2.0
